backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
pydantic
aiohttp
feedparser
beautifulsoup4
python-dotenv
transformers
torch
spacy
pandas
numpy
pytest
pytest-asyncio
httpx
redis
celery
sqlalchemy-utils
APScheduler
passlib[bcrypt]
python-jose[cryptography]
